refactor: move backend adapters to adapters folder
